@@ -2,23 +2,32 @@ apply plugin: 'com.android.application'
 
 apply plugin: 'kotlin-android'
 
-apply plugin: 'kotlin-android-extensions'
-
 android {
-    compileSdkVersion 29
+    namespace = "io.github.takusan23.tweetreader"
+    compileSdk 34
     defaultConfig {
         applicationId "io.github.takusan23.tweetreader"
-        minSdkVersion 24
-        targetSdkVersion 29
-        versionCode 6
-        versionName "3.1.0"
+        minSdk 21
+        targetSdk 34
+        versionCode 7
+        versionName "3.1.1"
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+    }
+    buildFeatures {
+        viewBinding true
     }
     buildTypes {
         release {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
+    }
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
+    kotlinOptions {
+        jvmTarget = "1.8"
     }
 }
 
@@ -30,7 +39,6 @@ dependencies {
     implementation 'org.twitter4j:twitter4j-core:4.0.7'
     implementation 'androidx.browser:browser:1.0.0'
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
     implementation "androidx.fragment:fragment:1.1.0"
     implementation "androidx.preference:preference:1.1.0"
     implementation 'androidx.appcompat:appcompat:1.1.0'
