@@ -1,18 +1,11 @@
 package io.github.takusan23.tweetreader
 
-import android.app.Activity
 import android.content.Intent
 import android.graphics.Bitmap
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import kotlinx.android.synthetic.main.activity_image_view.*
-import androidx.core.app.ComponentActivity.ExtraData
-import androidx.core.content.ContextCompat.getSystemService
-import android.icu.lang.UCharacter.GraphemeClusterBreak.T
 import android.widget.Toast
-import java.io.InputStream
-import java.net.URL
-
+import androidx.appcompat.app.AppCompatActivity
+import io.github.takusan23.tweetreader.databinding.ActivityImageViewBinding
 
 class ImageViewActivity : AppCompatActivity() {
 
@@ -23,36 +16,40 @@ class ImageViewActivity : AppCompatActivity() {
     var statusText = ""
     var pos = 0
 
+    private var _binding: ActivityImageViewBinding? = null
+    val binding get() = _binding!!
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_image_view)
+        _binding = ActivityImageViewBinding.inflate(layoutInflater)
+        setContentView(binding.root)
 
         supportActionBar?.hide()
 
-        val imageLinkList = intent.getStringArrayListExtra("images")
+        val imageLinkList = intent.getStringArrayListExtra("images")!!
         statusText = intent.getStringExtra("status") ?: "test"
         pos = intent.getIntExtra("pos", 0)
 
         //アニメーションに対応できるように
-        imageview_activity_view_pager.transitionName = "photo"
+        binding.imageviewActivityViewPager.transitionName = "photo"
 
         //ViewPagerによる切り替え機能
         val imageViewFragmentPagerAdapter =
             ImageViewFragmentPagerAdapter(supportFragmentManager, imageLinkList)
-        imageview_activity_view_pager.adapter = imageViewFragmentPagerAdapter
+        binding.imageviewActivityViewPager.adapter = imageViewFragmentPagerAdapter
 
         //位置反映
-        imageview_activity_view_pager.setCurrentItem(pos, true)
+        binding.imageviewActivityViewPager.setCurrentItem(pos, true)
 
         //閉じるボタン
-        imageview_activity_close_button.setOnClickListener {
+        binding.imageviewActivityCloseButton.setOnClickListener {
             finish()
         }
     }
 
     override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
         super.onActivityResult(requestCode, resultCode, data)
-        if (resultCode == Activity.RESULT_OK) {
+        if (resultCode == RESULT_OK) {
             if (requestCode == MediaSave.requestCode) {
                 val uri = data?.data
 
